Add support for 2-bay HDD enclosures. The first drive is for WiiU and the second drive is for vWii.
WiiU will recognize only the 1st drive but ignore the 2nd, which works in our favor - WiiU won't even nag you to format the 2nd drive to its proprietary format. usbtoggle is not required. Both drives are adequately powered by an external power brick because all 2-bay enclosures are externally powered. No more Y-cables. (WiiU is very strict about the 500mA USB current limit.)

If you have a RAID enclosure which can be set to RAID 0/1/JBOD/2LUN, set it to 2LUN. On a computer 2LUN simply shows as 2 separate hard drives.

This cIOS works only if the base IOS is either 56 or 57. base58 is not supported (because it does not patch the EHCI module).

Your homebrew shall remain on an SD card. The WiiU side demands that payload.elf be on an SD card and we can put all vWii homebrews on the same SD card, too. Most homebrews run with IOS 58 which does NOT see the second HDD.

In USB Loader GX, set USB Port to 1 to choose the second hard drive. You may have to restart for the change to take effect. Until now, the USB Port setting only works with an 'alt' cIOS such as d2x-v10-beta53-alt. The usual advice is to use non-alt cIOS and plug the hard drive to USB port 0. That advice still holds, the only change being that the same USB port 0 is now home to 2 HDDs, with the first hard drive appearing as 'USB Port 0' and the second hard drive appearing as 'USB Port 1'.

I have chosen to repurpose USB port selection as LUN selection so that no changes need to be made to USB Loader GX. The other solution considered was to scan LUNs in reverse order but that was rejected because I want no behavioural change compared to standard cIOS (d2x-v10-beta52) unless ordered by the user. Also considered was automatically detecting if the drive belongs to WiiU and choose the next drive if that is the case. That requires verifying some signature bytes on the disk and I do not wish to take risk on the initialization process.